chat-backend/requirements.txt
fastapi
groq
phidata
uvicorn
pydantic